Функция createCard перенесена в index.js
